Save test results

+++ src/com/gris/ege/activity/StartTestActivity.java
@@ -167,7 +167,7 @@ public class StartTestActivity extends Activity implements OnClickListener
                 {
                     case RESULT_START_TEST:
                     {
-                    	int aResultId=aData.getIntExtra(GlobalData.RESULT_ID, -1);
+                    	long aResultId=aData.getLongExtra(GlobalData.RESULT_ID, -1);
 
                     	Intent aCalculateIntent=new Intent();
                         aCalculateIntent.setClass(this, CalculateActivity.class);
